Update gpio_rain_service.py
Added ShutDown() process from StdService to close() the gpio pin in gpiozero when the Engine is shutdown by Weewx. WeeWX will shutdown the Engine if it loses a connection with the GW1000, and when the engine restarts it will reload the gpio_rain_service. Before this change, gpiozero would throw an exception because the pin was already assigned from before the Engine restarted.